% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardAnalysis.R
\name{analyzeassay.standard}
\alias{analyzeassay.standard}
\title{Quick Analysis}
\usage{
analyzeassay.standard(
  PathtoDatInfile,
  PathtoIds,
  GroupBy = c("DNA", "Protein", "Minutes"),
  BackgroundSignalType = Protein,
  NameofBackgroundSignal = "B",
  NormalizationSignalType = Protein,
  NameofNormalizationSignal = "C",
  SortGroup = "Protein",
  Writecsv = FALSE,
  Plot = TRUE,
  SavePlot = FALSE
)
}
\arguments{
\item{PathtoDatInfile}{A list of path names to the assay data to be imported.}

\item{PathtoIds}{A path to the file of identifiers for the assay data.}

\item{GroupBy}{The names of the variables over which the assay data is separated including minutes. Defaults to DNA, Protein and Minutes. Should not include "Replicate"}

\item{BackgroundSignalType}{The variable type of the signal to subtract from all other signals}

\item{NameofBackgroundSignal}{The quoted name of the signal to subtract from all other signals}

\item{NormalizationSignalType}{The variable type of the signal to divide all other signals}

\item{NameofNormalizationSignal}{The quoted name of the signal to divide all other signals}

\item{SortGroup}{The descriptive variable contained in the identifier file used to sort. Quoted}

\item{Writecsv}{A logical of whether to output a .csv in the current directory titled YYYYMMDD}
}
\value{
A report of the max normalized and background subtracted signal for the chosen group.
}
\description{
This method uses the \code{import.assaydat}, \code{dereplicate}, \code{subtract.background}, \code{normalize.bysignal} and \code{report.assaymax} functions to quickly process a standard dataset
}
