% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Central_Tendency.R
\name{Dereplicate}
\alias{Dereplicate}
\title{Dereplicate assay data}
\usage{
Dereplicate(
  AssayDat,
  GroupBy = c("DNA", "minutes", "protein"),
  Signal = intensity,
  Method = "CT"
)
}
\arguments{
\item{AssayDat}{Df of the long data table input file}

\item{GroupBy}{The names of the variables over which the assay data is separated including minutes. Defaults to DNA, protein and minutes. Should not include "replicate"}

\item{Signal}{The name of the assay signal being measured.}

\item{method}{Method used to dereplicate data. CT calculates mean, standard deviation and a 95\% confidence interval MM calculate max and min.}
}
\value{
A long kinetic data table with Mean, SD and a 95\% confidence internal of replicates
}
\description{
This function loads assay data as long data table with descriptive variables at each time point. And de-replicates them over the chosen variable.
}
