% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{find.maxsignal}
\alias{find.maxsignal}
\title{Find Biggest Signal}
\usage{
find.maxsignal(
  AssayDat,
  Signal = Signal,
  NormalizationSignalType = NULL,
  NameofNormalizationSignal = NULL
)
}
\arguments{
\item{AssayDat}{Df of the long data table input file}

\item{Signal}{Column name of signal}

\item{NormalizationSignalType}{The variable type of the signal to divide all other signals}

\item{NameofNormalizationSignal}{The quoted name of the signal to divide all other signals}
}
\value{
A vector, the max intensity of the normalization
}
\description{
Finds the maximum intensity of a single signal.
If NameofNormalizationSignal is Null, will return the max Signal for the enter Df.
}
