% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Normalize_kinetic.R
\name{find_biggest_signal}
\alias{find_biggest_signal}
\title{Find Biggest Signal}
\usage{
find_biggest_signal(kinetic_dataframe, mean_RFUs = "avg_RFU", Normal = NULL)
}
\arguments{
\item{kinetic_dataframe}{Data frame of}

\item{mean_RFUs}{Column name of signal}

\item{Normal}{Name of the normalization signal. Defaults to the name of the largest signal}
}
\value{
A vector, the max intensity of the normalization
}
\description{
Finds the maximum intensity of a single signal
}
