% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{import.assaydat}
\alias{import.assaydat}
\title{Import assay data and make long}
\usage{
import.assaydat(
  PathtoDatInfile,
  PathtoIds,
  ColstoExclude = c("Minutes"),
  By = "Well"
)
}
\arguments{
\item{PathtoDatInfile}{A list of path names to the assay data to be imported.}

\item{PathtoIds}{A path to the file of identifiers for the assay data.}

\item{ColstoExclude}{Pass into make.long.join Columns not made long.}

\item{By}{Pass into make.long.join. Key to join assay data and identifiers.}
}
\value{
A long assay data table of all of the input assay data bound to the identifying information provided by Ids.
}
\description{
Imports wide assay data from excel files and binds the identifying information. Returns long data. Defaults to kinetic data divided by well.
}
