% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_ImportFunctions.R
\name{import_kinetic}
\alias{import_kinetic}
\title{Import and make long}
\usage{
import_kinetic(excels = c("yymmdd_run_1A", "yymmdd_run_1B"), ids = identifiers)
}
\arguments{
\item{excels}{A list of path names to the kinetic data to be imported}

\item{id}{A path to the file of identifiers for the kinetic data}
}
\value{
A long kinetic data table of all of the input kinetic data bound to the identifying information provided by Ids.
}
\description{
Imports wide kinetic data from excel files and binds the identifying information. Returns long data.
}
