% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_ImportFunctions.R
\name{join_Txtl}
\alias{join_Txtl}
\title{Make long and join}
\usage{
join_Txtl(fl_dat, ids)
}
\arguments{
\item{fl_dat}{A wide data table of Assay data by well and time}

\item{ids}{A long data table of descriptive variables for each Assay well}
}
\value{
A data table of
}
\description{
Makes kinetic data long and joins with descriptive variables
}
