% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Normalize_kinetic.R
\name{normalize.byvalue}
\alias{normalize.byvalue}
\title{Make signal relative}
\usage{
normalize.byvalue(AssayDat, Signal = Signal, Value = MaxSignal)
}
\arguments{
\item{AssayDat}{A data frame containing long assay data}

\item{Signal}{Column name of the assay data to transform}

\item{Value}{A vector to divide the Signal by. Defaults to maximum value of selected Signal}
}
\value{
Df with intensity divided by the maximum intensity of the normalization signal
}
\description{
Divides all signals by a chosen signal to show relative intensity. Useful for normalizing over several experiments.
}
