% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Normalize_kinetic.R
\name{normalize_TxTl}
\alias{normalize_TxTl}
\title{Make signal relative}
\usage{
normalize_TxTl(
  kinetic_dataframe,
  mean_RFUs = avg_RFU,
  stnd_dev_RFUs = sd_RFU,
  max_RFU
)
}
\arguments{
\item{kinetic_dataframe}{A data frame containing kinetic data}

\item{mean_RFUs}{Name of the signal intensity}

\item{stnd_dev_RFUs}{Name of the standard deviation of the dereplicated signal intensity}

\item{max_RFU}{A vector, the max intensity of the normalization signal}
}
\value{
df with intensity divided by the maximum intensity of the normalization signal
}
\description{
Divides all signals by a chosen signal to show relative intensity. Useful for normalizing over several experiments.
}
