% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Normalize_kinetic.R
\name{normalizetoSignal}
\alias{normalizetoSignal}
\title{Normalize}
\usage{
normalizetoSignal(kdat)
}
\arguments{
\item{kdat}{long kinetic data table with central tendencies calculated}
}
\value{
a long kinetic data table with all values normalized to "Normal"
}
\description{
Divides all kinetic signals by a single kinetic signal. If the largest signal is chosen, makes all signals between 0-1.
}
