% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{report.assaymax}
\alias{report.assaymax}
\title{Report max assay data in a standard format}
\usage{
report.assaymax(
  AssayDat,
  GroupBy = c("Protein"),
  Signal = MeanSignal,
  SDSignal = SDSignal,
  SortBy = NULL,
  SortGroup = "Protein",
  Writecsv = FALSE
)
}
\arguments{
\item{AssayDat}{Df of the long data table input file}

\item{GroupBy}{The names of the variables over which the assay data is separated including minutes. Defaults to DNA, Protein and Minutes. Should not include "replicate"}

\item{Signal}{The name of the assay signal being measured.}

\item{SDSignal}{The name of the standard deviation of the assay signal.}

\item{SortBy}{The file path of the identifiers in Xlsx format used to sort the Report. Defaults to NULL.}

\item{SortGroup}{The descriptive variable contained in the identifier file used to sort. Quoted}

\item{Writecsv}{A logical of whether to output a .csv in the current directory titled YYYYMMDD}
}
\value{
A data table with Mean, SD and a 95\% confidence internal of replicates
}
\description{
This function loads assay data as long data table with descriptive variables at each time point
and reports the Maximum value with a confidence interval around that value.
}
