% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Central_Tendency.R
\name{stats_for_techreps}
\alias{stats_for_techreps}
\title{Compute measures of central tendency over technical replicates}
\usage{
stats_for_techreps(
  kdat,
  variables = c("DNA", "minutes", "protein"),
  signal = RFU
)
}
\arguments{
\item{kdat}{Path of the long kinetic data table input file}

\item{variables}{The names of the variables over which the assay data is separated including minutes. Defaults to DNA, protein and minutes. Should not include "replicate"}

\item{signal}{The name of the assay signal being measured. Defaults to RFU from Join_Txtl}
}
\value{
A long kinetic data table with Mean, SD and a 95\% confidence internal of replicates
}
\description{
This function loads kinetic data as long data table with descriptive variables at each timepoint.
}
