% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubtractBackground.R
\name{subtract.background}
\alias{subtract.background}
\title{Subtracts a background signal from all other signals}
\usage{
subtract.background(
  AssayDat,
  Signal = MeanSignal,
  GroupBy = c("Minutes"),
  BackgroundSignalType = DNA,
  NameofBackgroundSignal = "noDNA"
)
}
\arguments{
\item{AssayDat}{A df of Assay data}

\item{Signal}{The colname of the Assay intensity signal}

\item{GroupBy}{The names of the variables over which the assay data is separated. Defaults to minutes, cannot include BackgroundSignalType variable.}

\item{BackgroundSignalType}{The variable type of the signal to subtract from all other signals}

\item{NameofBackgroundSignal}{The quoted name of the signal to subtract from all other signals}
}
\value{
A df of Assay data with the intensity of the Background Signal subtracted
}
\description{
Subtracts a background signal from all other signals
}
