% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Background_subtraction.R
\name{subtract_background}
\alias{subtract_background}
\title{Subtracts a background signal from all other signals}
\usage{
subtract_background(
  AssayDat,
  Signal = mean_signal,
  GroupBy = c("minutes"),
  NameofBackgroundSignal = "noDNA",
  BackgroundSignalType = DNA
)
}
\arguments{
\item{AssayDat}{A df of Assay data}

\item{Signal}{The colname of the Assay intensity signal}

\item{GroupBy}{The names of the variables over which the assay data is separated. Defaults to minutes, cannot include BackgroundSignalType variable.}

\item{NameofBackgroundSignal}{The name of the signal to subtract from all other signals}

\item{BackgroundSignalType}{The variable type of the signal to subtract from all other signals}
}
\value{
A df of Assay data with the intensity of the Background Signal subtracted
}
\description{
Subtracts a background signal from all other signals
}
