% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_Background_subtraction.R
\name{subtract_background_kinetic}
\alias{subtract_background_kinetic}
\title{Subtracts a background signal from all other signals}
\usage{
subtract_background_kinetic(
  kdat,
  mean_signal = avg_RFU,
  NameofBackgroundSignal = "noDNA",
  groups = c("minutes")
)
}
\arguments{
\item{kdat}{A df of Assay data}

\item{mean_signal}{The colname of the Assay intensity signal}

\item{NameofBackgroundSignal}{The protein name of the signal to subtract from all other signals}
}
\value{
A df of Assay data with the intensity of the Background Signal subtracted
}
\description{
Subtracts a background signal from all other signals
}
