% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dereplicate.R
\name{dereplicate}
\alias{dereplicate}
\title{Dereplicate assay data}
\usage{
dereplicate(
  AssayDat,
  GroupBy = c("DNA", "Minutes", "Protein"),
  Signal = Intensity,
  Method = "CT"
)
}
\arguments{
\item{AssayDat}{Df of the long data table input file}

\item{GroupBy}{The names of the variables over which the assay data is separated including minutes. Defaults to DNA, Protein and Minutes. Should not include "Replicate"}

\item{Signal}{The name of the assay signal being measured.}

\item{Method}{Method used to dereplicate data. CT calculates mean, standard deviation and a 95\% confidence interval MM calculate max and min.}
}
\value{
A long kinetic data table with Mean, SD and a 95\% confidence internal of replicates
}
\description{
This function loads assay data as long data table with descriptive variables at each time point. And de-replicates them over the chosen variable.
}
