% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210716_ImportFunctions.R
\name{make.long.join}
\alias{make.long.join}
\title{Make long and join}
\usage{
make.long.join(AssayDat, Ids, ColstoExclude = c("minutes"), By = c("well"))
}
\arguments{
\item{AssayDat}{A wide data table of assay data by well and time.}

\item{Ids}{A long data table of descriptive variables for each assay well.}

\item{ColstoExclude}{Colnames of columns not to make long, defaults to "minutes" for kinetic data.}

\item{By}{Names of the location variable and key for joining identifiers to Assay data; defaults to "well".}
}
\value{
A data table of assay data bound to the descriptive variables in the identifiers file.
}
\description{
Makes assay data long and joins with descriptive variables.
}
